changed wordpress and mariadb

+++ srcs/requirements/wordpress/tools/config.sh
@@ -1,37 +1,28 @@
 #!/bin/bash
 
-# create directory to use in nginx container later and also to setup the wordpress conf
-mkdir /var/www/
-mkdir /var/www/html
+	sed -i "s/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/" "/etc/php/7.3/fpm/pool.d/www.conf";
+	chown -R www-data:www-data /var/www/*;
+	chown -R 755 /var/www/*;
+	mkdir -p /run/php/;
+	touch /run/php/php7.3-fpm.pid;
 
-cd /var/www/html
+if [ ! -f /var/www/html/wp-config.php ]; then
+	echo "Wordpress: setting up..."
+	mkdir -p /var/www/html
+	wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar;
+	chmod +x wp-cli.phar; 
+	mv wp-cli.phar /usr/local/bin/wp;
+	cd /var/www/html;
+	wp core download --allow-root;
+	mv /var/www/wp-config.php /var/www/html/
+	sed -i -r "s/db/$MYSQL_NAME/1"   /var/www/html/wp-config.php
+	sed -i -r "s/user/$MYSQL_USER/1"  /var/www/html/wp-config.php
+	sed -i -r "s/db_passwd/$MYSQL_PSW/1"    /var/www/html/wp-config.php
+	sed -i -r "s/localhost/mariadb/1"    /var/www/html/wp-config.php
+	echo "Wordpress: creating users..."
+	wp core install --allow-root --url=${DOMAIN_NAME} --title=${WP_TITLE} --admin_user=${WP_ADMIN_USER} --admin_password=${WP_ADMIN_PASSWORD} --admin_email=${WP_ADMIN_EMAIL}
+	wp user create --allow-root ${WP_USER} ${WP_USER_EMAIL} --user_pass=${WP_USER_PASSWORD};
+	echo "Wordpress: set up!"
+fi
 
-# remove all the wordpress files if there is something from the volumes to install it again
-rm -rf *
-
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
-
-chmod +x wp-cli.phar 
-
-mv wp-cli.phar /usr/local/bin/wp
-
-wp core download --allow-root
-mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-
-sed -i -r "s/db/$DB_NAME/1"   wp-config.php
-sed -i -r "s/user/$DB_USER/1"  wp-config.php
-sed -i -r "s/db_passwd/$DB_PASSWORD/1"    wp-config.php
-#sed -i -r "s/localhost/mariadb/1"    wp-config.php
-
-wp core install --url=$DOMAIN_NAME/ --title=$WP_TITLE --admin_user=$WP_ADMIN_USER --admin_password=$WP_ADMIN_PASSWORD --admin_email=$WP_ADMIN_EMAIL --skip-email --allow-root
-
-
-wp user create $WP_USER $WP_EMAIL --role=author --user_pass=$WP_USER_PASSWORD --allow-root
-
-wp theme install astra --activate --allow-root
-
-sed -i 's/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/g' /etc/php/7.3/fpm/pool.d/www.conf
-
-mkdir /run/php
-
-/usr/sbin/php-fpm7.3 -F
+exec "$@"
